refactor VerifiableCredentialsService introducing VerifiableCredentialTransformer
